Implement price input
The validation code that checks whether the input
is an positive integer divisible by the price of
one lotto ticket is also placed.

Special reformat code that deletes whitespaces
and commas are also added so that users can type
in numbers with commas for ease of use.

Also created new file "LottoConstants.java" that
stores the "global constants" of the project.
Currently only stores the cost of one ticket.